requirements.txt
PyGitHub==2.3.0
APScheduler==3.10.4
python-dotenv==1.0.1
